🎨 : rate limiting Redis, registrasi staff publik, dan transaksi stok-hutang atomik
Keamanan & akses:
- Tambah rate limiter (sliding window, 5 req/menit per IP) berbasis
  Redis storage; swagger UI "/" dan /storage dikecualikan
- CSRF dilewati untuk route /auth dan /api (API stateless berbasis Bearer)
- Aktifkan TrustProxy (loopback/private) agar c.IP() akurat di belakang Nginx
- Batasi endpoint sensitif (update/delete produk, delete transaksi/customer,
  report bulanan & hutang) hanya untuk superadmin via onlySuper

Role & registrasi:
- Hapus role "admin"; sisakan superadmin(0) dan staff(2)
- /auth/register kini publik dan selalu memaksa role "staff";
  admin/superadmin dibuat langsung lewat DB / CLI create-admin
- Role tanpa default DB agar superadmin (nilai 0) tidak ter-skip GORM saat insert

Transaksi atomik:
- CreateTransaction & DeleteTransaction kini berjalan dalam satu DB transaction:
  kurangi/kembalikan stok dengan row lock, upsert/balik saldo hutang, simpan/hapus
- Hitung subtotal & total di server (pakai SellingPriceDebt bila hutang);
  jangan percaya nilai dari client

Produk:
- Ubah Stock dari int ke float64 (DTO, domain, response, sheet, repo, usecase)
- Tambah unit "gram"; perbarui check constraint unit ke (0..5)

Perbaikan:
- Ganti uuid.MustParse/Must dengan uuid.Parse + error handling (hindari panic)
- Default order list ubah dari ASC ke DESC
- Fix Pluck "table not set" via .Model(&domain.Debts{}); tambah Debts.TableName()
- Perbaiki pesan error (transaction vs product) dan typo log

Dependency:
- Upgrade gofiber/fiber v3.1.0 -> v3.3.0; tambah gofiber/storage/redis v3

+++ internal/dto/request_dto/product_request.go
@@ -24,7 +24,7 @@ type AddProduct struct {
 	PurchasePrice    float64 `json:"purchase_price" validate:"required"`
 	SellingPrice     float64 `json:"selling_price" validate:"required"`
 	SellingPriceDebt float64 `json:"selling_price_debt" validate:"required"`
-	Stock            int     `json:"stock" validate:"required"`
+	Stock            float64 `json:"stock" validate:"required"`
 	Category         string  `json:"category" validate:"required"`
 	Image            string  `json:"image,omitempty"`
 }
@@ -53,8 +53,8 @@ type UpdateProduct struct {
 }
 
 type UpdateStock struct {
-	ID    string `json:"id" validate:"required"`
-	Stock int    `json:"stock,omitempty" validate:"required"`
+	ID    string  `json:"id" validate:"required"`
+	Stock float64 `json:"stock,omitempty" validate:"required"`
 }
 
 type GetAllProduct struct {
